# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PydanticObjectId": {
        "example": "5eb7cf5a86d9755df3a6c593",
        "maxLength": 24,
        "minLength": 24,
        "pattern": "^[0-9a-f]{24}$",
        "type": "string"
      },
...
      "RestoreMode": {
        "description": "Point-in-time treatment for objects created after the target.",
        "enum": [
          "exact",
          "non_destructive"
        ],
        "title": "RestoreMode",
        "type": "string"
      },
      "RestoreOperationListResponse": {
        "description": "Paginated restore operations.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RestoreOperationResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "RestoreOperationListResponse",
        "type": "object"
      },
...
      "RestorePlanRequest": {
        "description": "Selected immutable versions for a restore plan.",
        "properties": {
          "include_dependencies": {
            "default": true,
            "title": "Include Dependencies",
            "type": "boolean"
          },
          "mode": {
            "$ref": "#/components/schemas/RestoreMode",
            "default": "non_destructive"
          },
          "version_ids": {
            "items": {
              "$ref": "#/components/schemas/PydanticObjectId"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Version Ids",
            "type": "array"
          }
        },
        "required": [
          "version_ids"
        ],
        "title": "RestorePlanRequest",
        "type": "object"
      },
...
      "RestoreTargetListResponse": {
        "description": "Restore targets with the facet vocabularies the filter rows render.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RestoreTargetResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "sites": {
            "items": {
              "$ref": "#/components/schemas/RestoreTargetSiteResponse"
            },
            "title": "Sites",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "types": {
            "items": {
              "$ref": "#/components/schemas/RestoreTargetTypeCountResponse"
            },
            "title": "Types",
            "type": "array"
          }
        },
        "required": [
          "items",
          "total",
          "types",
          "sites"
        ],
        "title": "RestoreTargetListResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CsrfToken": {
        "description": "Double-submit CSRF token. Required on every unsafe method of a cookie session.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      },
      "SessionCookie": {
        "description": "Browser session issued by `POST /auth/login`. Unsafe methods must also send the `cg_csrf` cookie value in the `X-CSRF-Token` header.",
        "in": "cookie",
        "name": "cg_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
...
    "summary": "Configuration history, point-in-time recovery, and impact monitoring for Juniper Mist.",
    "title": "Mist Config Guardian",
    "version": "0.2.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/organizations/{organization_id}/restores": {
      "get": {
        "description": "List restore operations newest first.",
        "operationId": "list_restore_operations_api_v1_organizations__organization_id__restores_get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreOperationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "List Restore Operations",
        "tags": [
          "Restores"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/restores/plans": {
      "post": {
        "description": "Create a side-effect-free dependency-aware restore plan.",
        "operationId": "create_restore_plan_api_v1_organizations__organization_id__restores_plans_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestorePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreOperationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Create Restore Plan",
        "tags": [
          "Restores"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/restores/targets": {
      "get": {
        "description": "List restorable objects with the facet counts the filters render.",
        "operationId": "list_restore_targets_api_v1_organizations__organization_id__restores_targets_get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "all",
              "enum": [
                "all",
                "org",
                "site"
              ],
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "site_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site Id"
            }
          },
          {
            "in": "query",
            "name": "object_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Object Type"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreTargetListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "List Restore Targets",
        "tags": [
          "Restores"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/restores/{operation_id}": {
      "get": {
        "description": "Return one restore operation, including live per-action progress.",
        "operationId": "get_restore_operation_api_v1_organizations__organization_id__restores__operation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreOperationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Get Restore Operation",
        "tags": [
          "Restores"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/restores/{operation_id}/compensation": {
      "post": {
        "description": "Plan the reversal of every action a failed restore actually applied.",
        "operationId": "create_compensation_plan_api_v1_organizations__organization_id__restores__operation_id__compensation_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreOperationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Create Compensation Plan",
        "tags": [
          "Restores"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/restores/{operation_id}/compensation/execute": {
      "post": {
        "description": "Verify a fresh Mist administrator and queue the compensating plan.\n\nCompensation writes with the same delegated administrator identity a\nrestore does; it never falls back to the read-only service token.",
        "operationId": "execute_compensation_plan_api_v1_organizations__organization_id__restores__operation_id__compensation_execute_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreOperationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Execute Compensation Plan",
        "tags": [
          "Restores"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/restores/{operation_id}/execute": {
      "post": {
        "description": "Verify a fresh Mist administrator and queue the reviewed plan.",
        "operationId": "execute_restore_api_v1_organizations__organization_id__restores__operation_id__execute_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreOperationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Execute Restore",
        "tags": [
          "Restores"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/restores/{operation_id}/verification": {
      "get": {
        "description": "Return the post-restore checks recorded for one operation.",
        "operationId": "get_restore_verification_api_v1_organizations__organization_id__restores__operation_id__verification_get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreVerificationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Get Restore Verification",
        "tags": [
          "Restores"
        ]
      }
    },
...
  },
...
}